// src/app/shared/models/historial-producto.model.ts
// Generated by https://quicktype.io

export interface HistorialProducto {
    id:              number;
    idEmpresa:        number;
    idSucursal:      number;
    sucursal:        string;
    categoria:       string;
    idProducto:      number;
    codigoProducto:  string;
    producto:        string;
    imagen:          string;
    imagenRuta:      string;
    idTipoHistorial: number;
    tipoHistorial:   string;
    codigoSalida:   string;
    codigoStock:   string;
    descripcion:     string;
    entrada:         number;
    salida:          number;
    saldo:           number;
    precioCompra:    number;
    precioVenta:     number;
    precioVendido:     number;
    fecha:           string;
    hora:            string;
    usuario:         string;
    numeroSucursal:      number;
}